feat: roll driver 1.31.0-beta-1676591072000 (#1207)

+++ playwright/src/main/java/com/microsoft/playwright/impl/Protocol.java
@@ -94,7 +94,7 @@ class ExpectedTextValue {
 class FrameExpectOptions {
   Object expressionArg;
   List<ExpectedTextValue> expectedText;
-  Integer expectedNumber;
+  Double expectedNumber;
   SerializedArgument expectedValue;
   Boolean useInnerText;
   boolean isNot;
